% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_model.R
\docType{data}
\name{data_model}
\alias{data_model}
\title{Data-sample}
\format{A data.frame of time and precipitation in mm for every 3h interval.}
\usage{
data_model
}
\description{
EC-EARTH (GCM) rainfall data at the Gold Coast Seaway meteorologican station, Australia (station number - 40764, Period- 1/1/2000 to 12/31/2012, Latitude/longitude - -27.9390/153.4283)
}

