% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_model_future.R
\docType{data}
\name{data_model_future}
\alias{data_model_future}
\title{Data-sample}
\format{A data.frame of time and precipitation in mm for every 3h interval.}
\usage{
data_model_future
}
\description{
EC-EARTH (GCM) furure (RCP 4.5) rainfall data at the Gold Coast Seaway meteorologican station, Australia (station number - 40764, Period- 1/1/2026 to 12/31/2045, Latitude/longitude - -27.9390/153.4283)
}

